Adapt packages to use Protocol type class
The way ouroboros-network exposes Protocol information to clients such
as cardano-api and cardano-node was changed from a Protocol GADT,
which lived in ouroboros-network, to a type class that lives
in this repo; similarly for ProtocolClient. These changes make the
universe of Protocols that can be run more open than before.

cardano-api, cardano-cli, cardano-node, and cardano-node-chairman are
here modified to use this type class as much as possible. At some point
they do have to close their assumptions on which modes are supported, but
this is pushed up to the top-level as much as possible.